requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==6.1.0  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.1.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==3.2.12  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.49.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis

django-import-export==2.7.1
django-webpack-loader==1.4.1
djangorestframework==3.13.1
django-extensions==3.1.5  # https://github.com/django-extensions/django-extensions
tqdm==4.63.0
